Move SSR back to root directory
