+++ pages/1-preflight-preparation/C-xc-flight-planning.tsx
@@ -24,7 +24,7 @@ import {
   Warning,
 } from '../../components';
 import { ACS, uri } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const XcFlightPlanning: ACS.Page = (props) => {
