+++ pages/1-preflight-preparation/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { SectionPage } from '../../components';
 import { ACS } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const PreflightPreparation: ACS.Page = (props) => {
