+++ pages/2-preflight-procedures/A-ifr-systems.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { AIM, Bold, Collapse, Info, Link, Quotation, TaskPage, Warning } from '../../components';
 import { ACS, uri } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const IfrSystems: ACS.Page = (props) => {
