+++ pages/2-preflight-procedures/B-instruments.tsx
@@ -22,7 +22,7 @@ import {
   Warning,
 } from '../../components';
 import { ACS, uri } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const Instruments: ACS.Page = (props) => {
