+++ pages/2-preflight-procedures/C-flight-deck-check.tsx
@@ -19,7 +19,7 @@ import {
   Warning,
 } from '../../components';
 import { ACS, uri } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const FlightDeckCheck: ACS.Page = (props) => {
