+++ pages/3-atc-clearances-and-procedures/A-compliance.tsx
@@ -20,7 +20,7 @@ import {
   Warning,
 } from '../../components';
 import { ACS, uri } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const Compliance: ACS.Page = (props) => {
