+++ pages/3-atc-clearances-and-procedures/B-holding.tsx
@@ -15,7 +15,7 @@ import {
   Warning,
 } from '../../components';
 import { ACS } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const Holding: ACS.Page = (props) => {
