+++ pages/3-atc-clearances-and-procedures/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { SectionPage } from '../../components';
 import { ACS } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const ATCClearancesAndProcedures: ACS.Page = (props) => {
