+++ pages/4-instrument-flight/B-unusual-attitudes.tsx
@@ -13,7 +13,7 @@ import {
   Warning,
 } from '../../components';
 import { ACS, uri } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const UnusualAttitudes: ACS.Page = (props) => {
