+++ pages/5-navigation-systems/B-departure-enroute-arrival.tsx
@@ -18,7 +18,7 @@ import {
   Tooltip,
 } from '../../components';
 import { ACS, uri } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const DepartureEnrouteArrival: ACS.Page = (props) => {
