+++ pages/6-approach-procedures/B-precision.tsx
@@ -18,7 +18,7 @@ import {
   Warning,
 } from '../../components';
 import { ACS, uri } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const Precision: ACS.Page = (props) => {
