+++ pages/6-approach-procedures/D-circling.tsx
@@ -17,7 +17,7 @@ import {
   Warning,
 } from '../../components';
 import { ACS, uri } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const Circling: ACS.Page = (props) => {
