+++ pages/7-emergencies/A-loss-of-comm.tsx
@@ -16,7 +16,7 @@ import {
   Warning,
 } from '../../components';
 import { ACS, uri } from '../../lib';
-import { getStaticPropsFn } from '../../server/ssr';
+import { getStaticPropsFn } from '../../ssr';
 
 export const getStaticProps = getStaticPropsFn;
 const LossOfComm: ACS.Page = (props) => {
